increased precision for discretizers for regression and removed outer tree build routing redundancy

+++ cpp/src/BranchAssignmentObjectives/BranchAssignmentVariants.h
@@ -37,15 +37,15 @@ public:
 
 /** Regression MSE: statsDim = 2 (sum y, sum y²). */
 class SquaredErrorBranchAssignment
-    : public leaf_aggregate::LeafAggregationBranchAssignment<float> {
+    : public leaf_aggregate::LeafAggregationBranchAssignment<double> {
 public:
   static constexpr size_t kStatsDim = 2;
 
   SquaredErrorBranchAssignment(std::vector<size_t> &assignments,
                                size_t numPartitions,
-                               std::vector<std::vector<float>> &stats,
+                               std::vector<std::vector<double>> &stats,
                                std::vector<double> &leafWeights)
-      : leaf_aggregate::LeafAggregationBranchAssignment<float>(
+      : leaf_aggregate::LeafAggregationBranchAssignment<double>(
             assignments, numPartitions, stats, leafWeights, kStatsDim,
             std::make_unique<leaf_aggregate::SquaredErrorProcessor>()) {}
 };
