+++ cpp/src/Splitters/AbsoluteErrorSplitter.h
@@ -8,23 +8,23 @@
 #include "Splitter.h"
 #include "algorithms/WaveletTreeMAE.h"
 
-class AbsoluteErrorSplitter : public Splitter<float, float> {
+class AbsoluteErrorSplitter : public Splitter<double, float> {
 public:
   AbsoluteErrorSplitter(arma::frowvec &X, arma::frowvec &sampleWeights,
                         arma::Mat<float> &y);
   SplitCandidate makeRoot() override;
   float predict(const SplitCandidate &split) override;
   double score(const SplitCandidate &split) override;
-  double score(const std::vector<float> &stats, size_t l, size_t r) override;
-  const std::vector<float> &getStats(const SplitCandidate &split) override;
+  double score(const std::vector<double> &stats, size_t l, size_t r) override;
+  const std::vector<double> &getStats(const SplitCandidate &split) override;
 
 protected:
   const arma::Mat<float> &targets;
   WaveletTreeMAE waveletTree;
-  std::vector<float> empty_stats_cache_;
+  std::vector<double> empty_stats_cache_;
 
   QuantileStats getMedianQuantileStats(size_t l, size_t r);
 
-  void moveSample(std::vector<float> &rightStats, std::vector<float> &leftStats,
+  void moveSample(std::vector<double> &rightStats, std::vector<double> &leftStats,
                   size_t idx) override;
 };
